test: move lifecycle tests out of the plugin root
