+++ tests/accountIdentity.test.ts
@@ -1,13 +1,13 @@
 /*
  * OrionQuests account-identity regression tests.
  * Run from a Vencord checkout:
- * pnpm exec tsx --test src/userplugins/discord-quest-completer/accountIdentity.test.ts
+ * pnpm exec tsx --test src/userplugins/discord-quest-completer/tests/accountIdentity.test.ts
  */
 
 import assert from "node:assert/strict";
 import test from "node:test";
 
-import { isConfirmedDifferentAccount } from "./accountIdentity";
+import { isConfirmedDifferentAccount } from "../accountIdentity";
 
 test("a transient unknown identity is not an account change", () => {
     assert.equal(isConfirmedDifferentAccount(null, "user-a"), false);
